포인트 조회api

+++ src/main/java/com/cdy/ecommerce/ecommerce/admin/ProductAdminController.java
@@ -29,6 +29,6 @@ public class ProductAdminController {
 
     private ProductAdminDTO.Response entityToDTO(Product product) {
         // 변환
-        return ProductAdminDTO.Response.builder().id(product.getId()).name(product.getName()).price(product.getPrice()).build();
+        return ProductAdminDTO.Response.builder().id(product.getId()).name(product.getName()).price(product.getPrice().intValue()).build();
     }
 }
